Add a default points/required value in global settings
Just spent the past 45min debugging why the point values weren't being added. I thougth I might have messed something up because past versions of the code worked. Eventually, I realized that you need to have the question set to required = true for point values to be added... This is defintely going in the FAQ and the global settings page.