Allow each sheet to have different option lengths
Before, whatever option length on the global settings sheet applied to past created sheets. However, the option length could be different, which is why it is necessary to jot down the option length on each Sheet so that they unique value can be used on that sheet. Haven't fully tested this, but basic tests appear to be working