MY BRAIN WORKSgit add .!
I have figured out the Spreadsheet formula and I have pretty much figured out how the Javascript date object works. I have implemented the feature to populate dates between lessons, while accounting for the YRDSB holidays. This means that if there is a PA day on Friday, the code will write Monday (if there isn't a holiday) for the following lesson. THE HARDEST PART IS OVER!!!!!!